ai-visual-defect-detection-streamlit/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
Pillow>=10.2.0
streamlit>=1.32.0
